:up: data(float): change int to float input

+++ modulo3/lista-exercicios-3/questao1.py
@@ -4,7 +4,7 @@ seja inválido e continue pedindo até que o usuário informe um valor válido.
 '''
 
 while True:
-    nota = int(input('Digite uma nota de 0 a 10: '))
+    nota = float(input('Digite uma nota de 0 a 10: '))
     if nota < 0 or nota > 10:
         print('Valor inválido!')
     else:
